requirements.txt
langchain
langchain-community
langchain-mistralai
python-dotenv
faiss-cpu
langchain[mistralai]
streamlit